chore: 요청 응답 명세와 동일하도록 수정 및 개행 추가, 기타 로직 보완

+++ src/main/java/com/sixpack/dorundorun/feature/run/application/CreateRunSessionManualService.java
@@ -41,7 +41,7 @@ public class CreateRunSessionManualService {
 			runSession.getFinishedAt(),
 			runSession.getDurationTotal(),
 			runSession.getDistanceTotal(),
-			runSession.getPaceAvg(),
+			runSession.getPaceAvg().doubleValue(),
 			runSession.getCadenceAvg()
 		);
 	}
